+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-* Add `retry_time` parameter to retry for some period (in minutes) when an upload or download fails, instead of hardcoded 5 retries 
+* Add `retry_time` parameter to retry for some period (in seconds) when an upload or download fails, instead of hardcoded 5 retries 
 
 ### Fixed
 * Require `typing_extensions` on Python 3.11 (already required on earlier versinons) for better compatibility with pydantic v2